+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ CHANGELOG
 3.5
 ---
 
-* Removed HathiTrust Data API client code. Configurations for **HATHITRUST_OAUTH_KEY** and  **HATHITRUST_OAUTH_SECRET** are no longer needed in local settings.
+* HathiTrust Data API client code has been removed in favor of using rsync. Configurations for **HATHITRUST_OAUTH_KEY** and  **HATHITRUST_OAUTH_SECRET** are no longer needed in local settings.
 
 3.4
 ---